Se agrego un timer

+++ admin/eliminarpregunta.php
@@ -1,11 +1,16 @@
 <?php
+    session_start();
+
+    if (!isset($_SESSION['usuarioLogeado'])) {
+        header("Location:login.php");
+        exit;
+    }
 
     include("conexion.php");
     $id = $_GET['idPregunta'];
 
     $query = "DELETE FROM preguntas WHERE id = '$id'";
     mysqli_query($conn, $query);
+    header("Location:listadopreguntas.php");
+    exit;
 ?>
-<script>
-    window.location.href = 'listadopreguntas.php';
-</script>
